chore: move `as_context`
